refactoring user deletion to include confirmation before deletion (still under construction)

+++ src/users/delete-user/delete-user-service.ts
@@ -1,17 +1,25 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, ConflictException, BadRequestException } from '@nestjs/common';
 import { PrismaService } from 'src/database/PrismaService';
 
 @Injectable()
 export class DeleteUserService {
-  constructor(private readonly prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService) { }
 
-  async deleteUser(id: string): Promise<void> {
+  async deleteUser(id: string, confirmationPhrase: string): Promise<void> {
+
     try {
-      await this.prisma.user.delete({
-        where: {
-          id: parseInt(id),
-        },
-      });
+
+      if (!confirmationPhrase) {
+        throw new BadRequestException({ message: 'Para exclusão da conta é obrigatório inserir a frase de confirmação de exclusão' });
+
+      } else {
+
+        await this.prisma.user.delete({
+          where: {
+            id: parseInt(id),
+          },
+        });
+      }
     } catch (error) {
       throw new Error(`Erro ao excluir usuário: ${error.message}`);
     }
